Updated imports and fixed Blackboard
