+++ .idea/compiler.xml
@@ -8,8 +8,8 @@
         <outputRelativeToContentRoot value="true" />
         <module name="affectSimulator" />
         <module name="eyeSimulator" />
-        <module name="library" />
         <module name="headSimulatorTwo" />
+        <module name="library" />
         <module name="cobotSimulatorOne" />
         <module name="headSimulatorOne" />
         <module name="cobotSimulatorTwo" />
@@ -17,6 +17,7 @@
       </profile>
     </annotationProcessing>
     <bytecodeTargetLevel>
+      <module name="headSimTwo" target="20" />
       <module name="src" target="20" />
     </bytecodeTargetLevel>
   </component>
