+++ .idea/encodings.xml
@@ -15,6 +15,7 @@
     <file url="file://$PROJECT_DIR$/headSimulatorOne/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/headSimulatorTwo/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/headSimulatorTwo/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/library/headSimTwo/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/library/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/library/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
