+++ library/headSimTwo/pom.xml
@@ -4,30 +4,20 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-<!--    <modules>-->
-<!--        <module>headSimTwo</module>-->
-<!--    </modules>-->
-
-    <dependencies>
-        <dependency>
-            <groupId>org.eclipse.paho</groupId>
-            <artifactId>org.eclipse.paho.client.mqttv3</artifactId>
-            <version>1.2.5</version>
-        </dependency>
-    </dependencies>
+    <groupId>javiergs</groupId>
+    <artifactId>headSimTwo</artifactId>
+    <version>1.0-SNAPSHOT</version>
 
     <parent>
         <groupId>javiergs</groupId>
-        <artifactId>hcitoolbox</artifactId>
+        <artifactId>library</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-
-    <artifactId>library</artifactId>
 
     <properties>
         <maven.compiler.source>20</maven.compiler.source>
         <maven.compiler.target>20</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
-
+    
 </project>
