+++ library/src/main/java/headSim/TheSubscriber.java
@@ -1,5 +1,4 @@
-package app;
-
+package headSim;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
