+++ pom.xml
@@ -17,6 +17,7 @@
         <module>eyeSimulator</module>
         <module>headSimulatorTwo</module>
         <module>library</module>
+<!--        <module>library/headSimTwo</module>-->
     </modules>
 
     <properties>
